feat: make transfer network aware
