+++ services/local-notifications/core.go
@@ -11,7 +11,7 @@ import (
 	"github.com/ethereum/go-ethereum/p2p"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/status-im/status-go/multiaccounts/accounts"
-	"github.com/status-im/status-go/services/wallet"
+	"github.com/status-im/status-go/services/wallet/transfer"
 	"github.com/status-im/status-go/signal"
 )
 
@@ -91,13 +91,13 @@ type Service struct {
 	transmitter       *transmitter
 	walletTransmitter *transmitter
 	db                *Database
-	walletDB          *wallet.Database
+	walletDB          *transfer.Database
 	accountsDB        *accounts.Database
 }
 
 func NewService(appDB *sql.DB, chainID uint64) *Service {
 	db := NewDB(appDB, chainID)
-	walletDB := wallet.NewDB(appDB)
+	walletDB := transfer.NewDB(appDB)
 	accountsDB := accounts.NewDB(appDB)
 	trans := &transmitter{}
 	walletTrans := &transmitter{}
