+++ services/wallet/bigint/sql_big_int.go
@@ -1,4 +1,4 @@
-package wallet
+package bigint
 
 import (
 	"database/sql/driver"
