+++ services/wallet/transfer/balance_cache.go
@@ -1,4 +1,4 @@
-package wallet
+package transfer
 
 import (
 	"context"
