+++ services/wallet/transfer/database_test.go
@@ -1,4 +1,4 @@
-package wallet
+package transfer
 
 import (
 	"io/ioutil"
@@ -14,19 +14,19 @@ import (
 	"github.com/status-im/status-go/appdatabase"
 )
 
-func setupTestDB(t *testing.T) (*Database, func()) {
+func setupTestDB(t *testing.T) (*Database, *Block, func()) {
 	tmpfile, err := ioutil.TempFile("", "wallet-tests-")
 	require.NoError(t, err)
 	db, err := appdatabase.InitializeDB(tmpfile.Name(), "wallet-tests")
 	require.NoError(t, err)
-	return NewDB(db), func() {
+	return NewDB(db), &Block{db}, func() {
 		require.NoError(t, db.Close())
 		require.NoError(t, os.Remove(tmpfile.Name()))
 	}
 }
 
 func TestDBGetHeaderByNumber(t *testing.T) {
-	db, stop := setupTestDB(t)
+	db, _, stop := setupTestDB(t)
 	defer stop()
 	header := &types.Header{
 		Number:     big.NewInt(10),
@@ -40,7 +40,7 @@ func TestDBGetHeaderByNumber(t *testing.T) {
 }
 
 func TestDBGetHeaderByNumberNoRows(t *testing.T) {
-	db, stop := setupTestDB(t)
+	db, _, stop := setupTestDB(t)
 	defer stop()
 	rst, err := db.GetHeaderByNumber(777, big.NewInt(1))
 	require.NoError(t, err)
@@ -48,7 +48,7 @@ func TestDBGetHeaderByNumberNoRows(t *testing.T) {
 }
 
 func TestDBProcessBlocks(t *testing.T) {
-	db, stop := setupTestDB(t)
+	db, block, stop := setupTestDB(t)
 	defer stop()
 	address := common.Address{1}
 	from := big.NewInt(0)
@@ -70,7 +70,7 @@ func TestDBProcessBlocks(t *testing.T) {
 		Nonce:   &nonce,
 	}
 	require.NoError(t, db.ProcessBlocks(777, common.Address{1}, from, lastBlock, blocks))
-	t.Log(db.GetLastBlockByAddress(777, common.Address{1}, 40))
+	t.Log(block.GetLastBlockByAddress(777, common.Address{1}, 40))
 	transfers := []Transfer{
 		{
 			ID:          common.Hash{1},
@@ -86,7 +86,7 @@ func TestDBProcessBlocks(t *testing.T) {
 }
 
 func TestDBProcessTransfer(t *testing.T) {
-	db, stop := setupTestDB(t)
+	db, _, stop := setupTestDB(t)
 	defer stop()
 	header := &DBHeader{
 		Number:  big.NewInt(1),
@@ -116,7 +116,7 @@ func TestDBProcessTransfer(t *testing.T) {
 }
 
 func TestDBReorgTransfers(t *testing.T) {
-	db, stop := setupTestDB(t)
+	db, _, stop := setupTestDB(t)
 	defer stop()
 	rcpt := types.NewReceipt(nil, false, 100)
 	rcpt.Logs = []*types.Log{}
@@ -160,7 +160,7 @@ func TestDBReorgTransfers(t *testing.T) {
 }
 
 func TestDBGetTransfersFromBlock(t *testing.T) {
-	db, stop := setupTestDB(t)
+	db, _, stop := setupTestDB(t)
 	defer stop()
 	headers := []*DBHeader{}
 	transfers := []Transfer{}
@@ -201,167 +201,4 @@ func TestDBGetTransfersFromBlock(t *testing.T) {
 	require.NoError(t, err)
 	require.Len(t, rst, 4)
 
-}
-
-func TestGetNewRanges(t *testing.T) {
-	ranges := []*BlocksRange{
-		&BlocksRange{
-			from: big.NewInt(0),
-			to:   big.NewInt(10),
-		},
-		&BlocksRange{
-			from: big.NewInt(10),
-			to:   big.NewInt(20),
-		},
-	}
-
-	n, d := getNewRanges(ranges)
-	require.Equal(t, 1, len(n))
-	newRange := n[0]
-	require.Equal(t, int64(0), newRange.from.Int64())
-	require.Equal(t, int64(20), newRange.to.Int64())
-	require.Equal(t, 2, len(d))
-
-	ranges = []*BlocksRange{
-		&BlocksRange{
-			from: big.NewInt(0),
-			to:   big.NewInt(11),
-		},
-		&BlocksRange{
-			from: big.NewInt(10),
-			to:   big.NewInt(20),
-		},
-	}
-
-	n, d = getNewRanges(ranges)
-	require.Equal(t, 1, len(n))
-	newRange = n[0]
-	require.Equal(t, int64(0), newRange.from.Int64())
-	require.Equal(t, int64(20), newRange.to.Int64())
-	require.Equal(t, 2, len(d))
-
-	ranges = []*BlocksRange{
-		&BlocksRange{
-			from: big.NewInt(0),
-			to:   big.NewInt(20),
-		},
-		&BlocksRange{
-			from: big.NewInt(5),
-			to:   big.NewInt(15),
-		},
-	}
-
-	n, d = getNewRanges(ranges)
-	require.Equal(t, 1, len(n))
-	newRange = n[0]
-	require.Equal(t, int64(0), newRange.from.Int64())
-	require.Equal(t, int64(20), newRange.to.Int64())
-	require.Equal(t, 2, len(d))
-
-	ranges = []*BlocksRange{
-		&BlocksRange{
-			from: big.NewInt(5),
-			to:   big.NewInt(15),
-		},
-		&BlocksRange{
-			from: big.NewInt(5),
-			to:   big.NewInt(20),
-		},
-	}
-
-	n, d = getNewRanges(ranges)
-	require.Equal(t, 1, len(n))
-	newRange = n[0]
-	require.Equal(t, int64(5), newRange.from.Int64())
-	require.Equal(t, int64(20), newRange.to.Int64())
-	require.Equal(t, 2, len(d))
-
-	ranges = []*BlocksRange{
-		&BlocksRange{
-			from: big.NewInt(5),
-			to:   big.NewInt(10),
-		},
-		&BlocksRange{
-			from: big.NewInt(15),
-			to:   big.NewInt(20),
-		},
-	}
-
-	n, d = getNewRanges(ranges)
-	require.Equal(t, 0, len(n))
-	require.Equal(t, 0, len(d))
-
-	ranges = []*BlocksRange{
-		&BlocksRange{
-			from: big.NewInt(0),
-			to:   big.NewInt(10),
-		},
-		&BlocksRange{
-			from: big.NewInt(10),
-			to:   big.NewInt(20),
-		},
-		&BlocksRange{
-			from: big.NewInt(30),
-			to:   big.NewInt(40),
-		},
-	}
-
-	n, d = getNewRanges(ranges)
-	require.Equal(t, 1, len(n))
-	newRange = n[0]
-	require.Equal(t, int64(0), newRange.from.Int64())
-	require.Equal(t, int64(20), newRange.to.Int64())
-	require.Equal(t, 2, len(d))
-
-	ranges = []*BlocksRange{
-		&BlocksRange{
-			from: big.NewInt(0),
-			to:   big.NewInt(10),
-		},
-		&BlocksRange{
-			from: big.NewInt(10),
-			to:   big.NewInt(20),
-		},
-		&BlocksRange{
-			from: big.NewInt(30),
-			to:   big.NewInt(40),
-		},
-		&BlocksRange{
-			from: big.NewInt(40),
-			to:   big.NewInt(50),
-		},
-	}
-
-	n, d = getNewRanges(ranges)
-	require.Equal(t, 2, len(n))
-	newRange = n[0]
-	require.Equal(t, int64(0), newRange.from.Int64())
-	require.Equal(t, int64(20), newRange.to.Int64())
-	newRange = n[1]
-	require.Equal(t, int64(30), newRange.from.Int64())
-	require.Equal(t, int64(50), newRange.to.Int64())
-	require.Equal(t, 4, len(d))
-}
-
-func TestInsertRange(t *testing.T) {
-	db, stop := setupTestDB(t)
-	defer stop()
-
-	r := &BlocksRange{
-		from: big.NewInt(0),
-		to:   big.NewInt(10),
-	}
-	nonce := uint64(199)
-	balance := big.NewInt(7657)
-	account := common.Address{2}
-
-	err := db.InsertRange(777, account, r.from, r.to, balance, nonce)
-	require.NoError(t, err)
-
-	block, err := db.GetLastKnownBlockByAddress(777, account)
-	require.NoError(t, err)
-
-	require.Equal(t, 0, block.Number.Cmp(r.to))
-	require.Equal(t, 0, block.Balance.Cmp(balance))
-	require.Equal(t, nonce, uint64(*block.Nonce))
 }
