+++ services/wallet/transfer/transmitter.go
@@ -1,4 +1,4 @@
-package wallet
+package transfer
 
 import (
 	"sync"
